remove FnDiverges in favor of FnTypeDiverges (#13)
We'll want to be able to define and run functions that trivially diverge,
such as a TCP server. Eventually we'll add a bottom type and require
annotating the output as such. For now, we infer `(∀ v0 v1 . v0 -> v1)`.

However, we still want to ban functions that do not have a stable inferred
type. For those, we return `FnTypeDiverges`.