add failing inferType test
When composing an undefined function's type, missing stacks should
not be added as (v0 -> v0). This general approach seems to be a
dead end. We need for inferType to fail if it needs a type
for an undefined function. If it can infer the type via
fully defined match cases, then it does not need to fail, though.
We'll just need to add a second `checkType` pass.