Fix inferType and defineFns (#20)
* add failing inferType test

When composing an undefined function's type, missing stacks should
not be added as (v0 -> v0). This general approach seems to be a
dead end. We need for inferType to fail if it needs a type
for an undefined function. If it can infer the type via
fully defined match cases, then it does not need to fail, though.
We'll just need to add a second `checkType` pass.

* remove `Result` type alias

* inferType: error on undefined function

- throw UndefinedFn on ECall of undefined function
- throw UndefinedFn if all match cases throw UndefinedFn
- otherwise, filter out any match cases that throw UndefinedFn

* add checkType; rework defineFns

* remove undefinedFnIds

* checkType: use `match`, not `mgu`

* fix defineFns, add tests

* add defineFns test pair, with one failing

Depending on the dependency sorting implementation, one of these tests will
fail and one will succeed. This is problematic.

* alter failing test to pass

thereby restricting the implementation of dependencySortFns